move : artist/manager 파일 분리
